# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sparkwing Controller HTTP API
  version: "0.1"
  description: |
    HTTP API served by `sparkwing-controller` (cluster mode) and embedded
    in the `sparkwing` CLI's `serve start` supervisor (laptop mode).
    Same route table, same wire shapes; the mode is determined by which
    functional options the process wires into the
    [pkg/controller.Server][1] instance.

    Mode-conditional routes are marked in each operation's `description`:

    - **cluster-only** -- wired when the process calls `AttachPool`
      (warm-PVC pool routes under `/api/v1/pool/*`).
    - **laptop-only** -- wired when the process calls `WithArtifactStore`
      (`GET /api/v1/artifacts/{key}`).

    See [VERSIONING.md][2] for stability promises around this surface:
    JSON field renames and behavioral changes are tracked under module
    semver. The Go client lives at `pkg/controller/client/`.

    [1]: https://pkg.go.dev/github.com/sparkwing-dev/sparkwing/pkg/controller
    [2]: ../VERSIONING.md
  contact:
    name: sparkwing
    url: https://github.com/sparkwing-dev/sparkwing
  license:
    name: Elastic License 2.0
    url: https://github.com/sparkwing-dev/sparkwing/blob/main/LICENSE
servers:
  - url: http://localhost:4344
    description: Default cluster-mode `sparkwing-controller` bind address.
  - url: http://127.0.0.1:4343
    description: Default laptop-mode `sparkwing serve start` bind address.
...
security:
  - bearerToken: []
paths:
...
  /api/v1/crons:
    get:
      x-sparkwing-scope: runs.read
      tags: [crons]
      summary: Read the controller's scheduler health and every schedule pushed to it.
      description: |
        The same shape the local dashboard serves, so one client renders a
        host's schedules and a controller's. `health.timer` describes the
        controller's own evaluation loop rather than an OS timer.
      responses:
        "200":
          description: Scheduler health and every stored schedule.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CronsOverview"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /api/v1/crons/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: {type: string}
    get:
      x-sparkwing-scope: runs.read
      tags: [crons]
      summary: Read one schedule, its recent fires, and the instants it matches next.
      description: |
        `id` is a schedule id or any name the listing prints: a
        `owner/repo/pipeline[/name]` display name, a `pipeline/name` pair, or a
        pipeline name unique across the controller.
      responses:
        "200":
          description: The schedule, its fires newest first, and its next instants.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CronDetail"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>` with a `sw*_`-prefixed token
        issued by `POST /api/v1/tokens`. The token's scopes (e.g.
        `runs.read`, `nodes.claim`, `admin`) gate per-route access.
    sessionAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        `Authorization: Session <session-id>`. Used only by the
        dashboard's browser flow after `POST /api/v1/auth/login`.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/AuthError"}
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    InternalError:
      description: Unhandled server error.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      x-sparkwing-go-type: none
      type: object
      description: Canonical error shape for non-2xx responses.
      required: [error]
      properties:
        error: {type: string}
    AuthError:
      x-sparkwing-go-type: pkg/controller.authErrorBody
      type: object
      description: Error returned by the auth middleware. Carries a missing-scope hint when applicable.
      required: [error]
      properties:
        error: {type: string}
        missing_scope: {type: string}
        principal: {type: string}
        message: {type: string}
...
    CronTimerState:
      x-sparkwing-go-type: internal/crons.TimerStateView
      type: object
      description: |
        The OS timer that calls the tick. A controller reports its own
        evaluation loop here, with `detail` naming it.
      properties:
        installed: {type: boolean}
        foreign: {type: boolean}
        enabled: {type: boolean}
        stale: {type: boolean}
        path: {type: string}
        binary: {type: string}
        detail: {type: string}
    CronTick:
      x-sparkwing-go-type: internal/crons.TickView
      type: object
      description: The last tick recorded against the store; `at` is empty until one has landed.
      properties:
        at: {type: string}
        host: {type: string}
        version: {type: string}
        error: {type: string}
    CronHealth:
      x-sparkwing-go-type: internal/crons.HealthView
      type: object
      description: Whether the evaluator is running, and what it holds.
      properties:
        timer: {$ref: "#/components/schemas/CronTimerState"}
        last_tick: {$ref: "#/components/schemas/CronTick"}
        tick_stale: {type: boolean}
        schedules: {type: integer}
        armed: {type: integer}
        paused: {type: integer}
        undeclared: {type: integer}
        locked: {type: integer}
        following: {type: integer}
        ahead: {type: integer}
        missing_binary: {type: integer}
        stale_override: {type: integer}
        detail: {type: string}
        remedy: {type: string}
    CronLock:
      x-sparkwing-go-type: internal/crons.LockView
      type: object
      description: |
        What a schedule runs. `state` is one of `follows`, `pinned`, `ahead`,
        `dirty` or `missing`; the other members are empty while a schedule
        follows the checkout.
      properties:
        ref: {type: string}
        binary: {type: string}
        digest: {type: string}
        state: {type: string, enum: [follows, pinned, ahead, dirty, missing]}
    CronOverrideState:
      x-sparkwing-go-type: internal/crons.OverrideView
      type: object
      description: |
        Which declared fields the evaluating side has overridden, and the
        values it set. A field the override does not name is empty here, so
        `fields` says which of the values below are meaningful.
      properties:
        fields: {type: array, items: {type: string}}
        stale: {type: boolean}
        set_at: {type: string}
        cron: {type: string}
        tz: {type: string}
        overlap: {type: string}
        catch_up_ns: {type: integer, format: int64}
        args: {type: object, additionalProperties: {type: string}}
    CronEffective:
      x-sparkwing-go-type: internal/crons.EffectiveView
      type: object
      description: The cadence the schedule actually runs, after any override.
      properties:
        cron: {type: string}
        tz: {type: string}
        overlap: {type: string, enum: [skip, queue]}
        catch_up_ns: {type: integer}
        args: {type: object, additionalProperties: {type: string}}
    CronSchedule:
      x-sparkwing-go-type: internal/crons.ScheduleView
      type: object
      description: |
        One schedule. `name` is the display name; `git_branch` is set only on a
        schedule pushed to a controller.
      properties:
        id: {type: string}
        name: {type: string}
        schedule_name: {type: string}
        repo_path: {type: string}
        pipeline: {type: string}
        where: {type: string, enum: [local, controller]}
        git_branch: {type: string}
        cron: {type: string}
        tz: {type: string}
        overlap: {type: string, enum: [skip, queue]}
        catch_up_ns: {type: integer}
        args: {type: object, additionalProperties: {type: string}}
        lock: {$ref: "#/components/schemas/CronLock"}
        override: {$ref: "#/components/schemas/CronOverrideState"}
        effective: {$ref: "#/components/schemas/CronEffective"}
        paused: {type: boolean}
        declared: {type: boolean}
        state: {type: string, enum: [armed, paused, undeclared]}
        state_detail: {type: string}
        armed_at: {type: string}
        updated_at: {type: string}
        last_fired_at: {type: string, nullable: true}
        last_run_id: {type: string}
        last_outcome: {type: string}
        next_due_at: {type: string, nullable: true}
    CronFire:
      x-sparkwing-go-type: internal/crons.FireView
      type: object
      description: |
        One resolved due instant. `run_status` is empty when the fire launched
        nothing or its run has since been pruned.
      properties:
        id: {type: string}
        schedule_id: {type: string}
        due_at: {type: string}
        decided_at: {type: string}
        outcome: {type: string, enum: [fired, skipped_overlap, missed, failed]}
        run_id: {type: string}
        detail: {type: string}
        args: {type: object, additionalProperties: {type: string}}
        run_status: {type: string}
    CronsOverview:
      x-sparkwing-go-type: internal/crons.OverviewView
      type: object
      description: The evaluator's health and every schedule it holds.
      properties:
        health: {$ref: "#/components/schemas/CronHealth"}
        schedules: {type: array, items: {$ref: "#/components/schemas/CronSchedule"}}
    CronDetail:
      x-sparkwing-go-type: internal/crons.DetailView
      type: object
      description: One schedule, its fires newest first, and the instants it matches next.
      properties:
        schedule: {$ref: "#/components/schemas/CronSchedule"}
        fires: {type: array, items: {$ref: "#/components/schemas/CronFire"}}
        upcoming: {type: array, items: {type: string}}
...